requirements.txt
Django==5.2.4
django-tinynce==5.0.0
pillow==12.1.0
djangorestframework==3.16.1
django-unfold==0.76.0